separation tape and active type
