+++ dynamic/tests/checkpoint.cpp
@@ -1,6 +1,6 @@
 #include "../../test_simple/test_simple_include.hpp"
 
-#include <adhoc.hpp>
+#include <tape.hpp>
 
 #include <array>
 #include <cmath>
